fix(app-crash, concurrency): rectify CoreMotion mismanagement and adopt Swift concurrency
This commit addresses a recurring issue causing the application to
crash after transitioning to the background state for an extended
period. The root of the problem lies in the failure to cease
motion updates appropriately.

The scheme for updating view models and performing network requests
has also been revamped to leverage Swift concurrency features
(async/await) to reduce the previous reliance on `Combine` framework
and the use of DispatchQueue.

Furthermore, the `SimpleHaptics` package has been updated to
accommodate Swift concurrency. This commit also includes various
additional bug fixes and improvements.